Uday + Bhavya Chnges

+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Events;
+use App\Models\Order;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -15,7 +16,7 @@ class UserController extends Controller
     {
         $user = Auth::user();
         if ($user && $user->id === 0) {
-            $users = User::all();
+            $users = User::paginate(10);
             return view('Admin.UserManagement', compact('users'));
         } else {
             abort(403, 'Unauthorized');
@@ -23,13 +24,14 @@ class UserController extends Controller
     }
 
     // show all the ticket details of the all the users.
-    public function tickets()
+    public function tickets($id)
     {
         $user = Auth::user();
         if ($user && $user->id === 0) {
 
-            $events = Events::all();
-            return view('Admin.UserTicketManagement', compact('events'));
+            $username = User::where('id', $id)->first(["name"])->name;
+            $events = Events::where('organizer_id', $id)->orderByDesc('created_at')->paginate(5);
+            return view('Admin.viewEventsByUserId', compact('events', 'username'));
         } else {
             abort(403, 'Unauthorized');
         }
@@ -39,12 +41,19 @@ class UserController extends Controller
     // Delete the any user.
     public function destroy($id)
     {
-        try {
+
+        $hasEvents = Events::where('organizer_id', $id)->first();
+        if ($hasEvents) {
+            return redirect()->back()->with('error', 'The user has organized events and cannot be deleted.');
+        }
+        $hasOrders = Order::where('user_id', $id)->first();
+        if ($hasOrders) {
+            return redirect()->back()->with('error', 'The user has orders as an organizer and cannot be deleted.');
+        } else {
             $user = User::findOrFail($id);
             $user->delete();
-            return response()->json(['message' => 'User deleted successfully']);
-        } catch (\Exception $e) {
-            return response()->json(['message' => 'Error deleting user: ' . $e->getMessage()], 500);
+            // return response()->json(['su' => 'User deleted successfully'])
+            return redirect()->back()->with('success', 'User Deleted Succefully.');
         }
     }
 }
